ticket-to-pkpass-starter-python/requirements.txt
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.13
pyzbar==0.1.9
python-dateutil==2.9.0.post0
dateparser==1.2.0
jsonschema==4.23.0
click==8.1.7
openai==1.40.2
